split views_model.py into modules by class
